mt5_backend/requirements.txt
django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.3
MetaTrader5>=5.0.45
python-dotenv>=1.0
